requirements.txt
python-dotenv==1.0.1
aiogram==3.26.0
aiosqlite==0.19.0
validators==0.22.0
sqlalchemy==2.0.47